Fix init state selector memoization
Replace createShallowArrayCompareSelector with
createShallowEqualSelector, which can handle both object and
array equality checks.

Wrap the selector in a new selector that does a shallow equality
check. This will ensure that the selector does not recompute its
output when state.selfInit or state.prepared changes in keys that
are not related to the current prepareKey