Create environment for integration tests
Create a IsomorphicTestEnvironment utility that can be used in
integration tests to simulate a component rendering on the
server first and then on the client. This can be used to
test more complex situations.

Create 2 sample integration tests using the new utility. This
file is named .test.js to distinguish from the unit tests, which
are named .spec.js

Create an ExtendableError utility to create custom error classes.
Use this to create a custom error type PrepareValidationError. This
can be used to assert that a prepare validation error has occured
in tests (as opposed to another type of error).

Add the babel-plugin-transform-builtin-extend plugin to make sure
we can extend the builtin Error correctly.