Allow passing a custom component id
Allow to pass a string argument instead of a component to
withInitAction, which will return a new wrapper function that
wraps a component with a custom component id. This can be
used whenever displayName does not suffice to provide unique
component ids.

This new parameter is passed in a curried way to conform to
the conventions of HoC but still be backwards compatible.